+++ internal/pkg/config/errors.go
@@ -1,5 +1,0 @@
-package config
-
-import "errors"
-
-var ErrGHUserRequired = errors.New("GITRAKZ_GH_USER is required")